Move decompressor into dedicated package
